chore: refine startup check for test MQ containers
The integration test containers waits for the queue manager to be
ready before starting.

The string that was being used does show up in the log earlier
than expected (when the queue manager logs out a list of error /
message codes) so this commit changes it to something that is
more representative of the queue manager being ready to go.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>