feat: configurable retry interval after retriable exceptions
If the connector throws a retriable exception, there is no need
for Connect's Kafka consumer to wait for as long to re-consume
messages from the Kafka topic. It's therefore helpful to be able
to specify that the consumer should use a shorter timeout for
polls immediately following a retriable exception.

This commit introduces a new config parameter mq.retry.backoff.ms
which allows this to be controlled.

By default, it is set to 1 minute, which is how long it
currently waits.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>